Lots of fixes including double-buffer rendering, x- and y-flip, a fix to ADC, the half-carry flag, and opcode logging.
Lots of fixes including: double-buffer rendering to prevent hangs on close (the LCDControl reads from the front buffer at ~60fps); x- and y-flip support added to the PPU's tile rendering; a dumb typo in the ADC opcode; more correct setting of the half-carry flag, and logging of opcodes and CPU state to a file. The emulator is more accurate, but slower now.