SampleProviders now use a ring buffer for writing and reading audio data.
SampleProviders now use a ring buffer for writing and reading audio data. Each frame, waveform data is written to the audio buffer and NAudio reads this data as required. Music and SFX are now consistently on-beat, but there are timing issues still with NAudio not quite getting the data it needs.